+++ src/Modules/Profile/Profile.js
@@ -11,6 +11,7 @@ import MediationRequest from './components/MediationRequest'
 import ReportRequest from './components/ReportRequest'
 import BananaList from './components/BananaList'
 import PublicProfile from './components/PublicProfile'
+import ProfileStoreList from './components/ProfileStoreList'
 import PickupHistory from '../StoreUser/components/PickupHistory'
 // Wallpost
 import { URL_PART } from '@/browser'
@@ -32,6 +33,7 @@ async function trySendBuddyRequest (userId) {
 vueRegister({
   BananaList,
   PickupHistory,
+  ProfileStoreList,
   PublicProfile,
   MediationRequest,
   ReportRequest,
@@ -39,6 +41,7 @@ vueRegister({
 
 vueApply('#vue-profile-bananalist', true) // BananaList
 vueApply('#vue-pickup-history', true) // PickupHistory
+vueApply('#vue-profile-storelist', true) // ProfileStoreList
 vueApply('#profile-public', true) // PublicProfile
 vueApply('#mediation-Request', true) // MediationRequest
 vueApply('#report-Request', true) // ReportRequest